commit for seci=urity reason. bug fixing on sequence diagram in progress
git-svn-id: https://violet.svn.sourceforge.net/svnroot/violet@71 a3dc6853-a221-45bc-9d17-2c9bfbe3703c

+++ VioletFramework/src/main/java/com/horstmann/violet/product/diagram/abstracts/node/AbstractNode.java
@@ -401,7 +401,7 @@ public abstract class AbstractNode implements INode
     
     private static final Color SHADOW_COLOR = Color.LIGHT_GRAY;
     protected static Color BACKGROUND_COLOR = UIManager.getColor("TextPane.background");
-    public static final int SHADOW_GAP = 4;
+    public static final double SHADOW_GAP = 4;
 
     private ArrayList<INode> children;
     private INode parent;
